requirements.txt
# Gabriel — Dependencies
# Install: pip install -r requirements.txt

# Core — Camera & Frame Processing
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
ImageHash>=4.3.0

# Vision API
google-generativeai>=0.5.0

# Chat API
anthropic>=0.25.0

# API Server
fastapi>=0.110.0
uvicorn>=0.27.0

# Alerts
python-telegram-bot>=21.0

# Utilities
python-dotenv>=1.0.0
